Renamed some api classes

+++ src/main/java/dev/lazurite/rayon/impl/physics/manager/DebugManager.java
@@ -94,7 +94,7 @@ public final class DebugManager {
                             RenderSystem.depthMask(false);
                             RenderSystem.lineWidth(1.0F);
 
-                            FloatBuffer buffer = DebugShapeFactory.getDebugTriangles(body.getCollisionShape(), 0).rewind();
+                            FloatBuffer buffer = (FloatBuffer) DebugShapeFactory.getDebugTriangles(body.getCollisionShape(), 0).rewind();
                             BufferBuilder builder = Tessellator.getInstance().getBuffer();
                             float alpha = ((DebuggableBody) body).getOutlineAlpha();
                             Vector3f color = ((DebuggableBody) body).getOutlineColor();
